removed RTC from cubemx, made system debuggable on Wokwi simulator, fixed ADC value calculation for temperature sensor

+++ nucleo_c031c6/Core/Inc/state_machine.h
@@ -17,7 +17,7 @@ typedef enum {
     SYSTEM_EMERGENCY_STOP
 } SystemState;
 
-SystemState updateSystemState(float temperature, float threshold, float tolerance, 
+SystemState updateSystemState(float temperature, uint8_t threshold, uint8_t tolerance, 
                             uint8_t sensor_fault, uint8_t emergency_stop);
 
 #ifdef __cplusplus
