+++ nucleo_c031c6/Core/Src/state_machine.cpp
@@ -1,6 +1,6 @@
 #include "state_machine.h"
 
-SystemState updateSystemState(float temperature, float threshold, float tolerance, 
+SystemState updateSystemState(float temperature, uint8_t threshold, uint8_t tolerance, 
                             uint8_t sensor_fault, uint8_t emergency_stop) {
 
         if(emergency_stop) {
